Feature: bind models at api controller; p.s. nth changed

+++ MathEvent/MathEvent/Helpers/File/FileUpload.cs
@@ -27,11 +27,14 @@ namespace MathEvent.Helpers.File
             {
                 if (file != null)
                 {
-                    var imageToBeDeleted = UserDataPathWorker.GetRootPath(Path.Combine(performance.DataPath, performance.PosterName));
-
-                    if (System.IO.File.Exists(imageToBeDeleted))
+                    if (performance.PosterName != null)
                     {
-                        System.IO.File.Delete(imageToBeDeleted);
+                        var imageToBeDeleted = UserDataPathWorker.GetRootPath(Path.Combine(performance.DataPath, performance.PosterName));
+
+                        if (System.IO.File.Exists(imageToBeDeleted))
+                        {
+                            System.IO.File.Delete(imageToBeDeleted);
+                        }
                     }
 
                     var fileName = file.Name;
